+++ test/unit/render/denoise/BilateralDenoiserTest.cpp
@@ -40,6 +40,28 @@ namespace BilateralDenoiserTest {
     ASSERT_COLOR_NEAR(Colord::white(), buffer[0][2], 1e-6);
   }
 
+  TEST(BilateralDenoiser, UsesFeatureBuffersToPreserveGeometryEdges) {
+    render::BilateralDenoiser denoiser(1, 10.0);
+    Buffer<Colord> beauty(3, 1);
+    beauty[0][0] = Colord::black();
+    beauty[0][1] = Colord(3.0, 0.0, 0.0);
+    beauty[0][2] = Colord::black();
+
+    Buffer<Colord> albedo(3, 1);
+    albedo[0][0] = Colord::black();
+    albedo[0][1] = Colord::white();
+    albedo[0][2] = Colord::black();
+
+    render::DenoiserFrame frame(beauty);
+    frame.features.albedo = &albedo;
+
+    denoiser.denoiseFrame(frame);
+
+    EXPECT_LT(beauty[0][0][0], 0.01);
+    EXPECT_GT(beauty[0][1][0], 2.99);
+    EXPECT_LT(beauty[0][2][0], 0.01);
+  }
+
   TEST(BilateralDenoiser, ClonesConfiguration) {
     render::BilateralDenoiser denoiser(3, 0.25);